VarCorr {ordinal}	R Documentation

Extract variance and correlation parameters

Description

The VarCorr function extracts the variance and (if present) correlation parameters for random effect terms in a cumulative link mixed model (CLMM) fitted with clmm.

Usage

## S3 method for class 'clmm'
VarCorr(x, ...)

Arguments

x
a clmm object.

...
currently not used by the clmm method.

Details

The VarCorr method returns a list of data.frames; one for each distinct grouping factor. Each data.frame has as many rows as there are levels for that grouping factor and as many columns as there are random effects for each level. For example a model can contain a random intercept (one column) or a random intercept and a random slope (two columns) for the same grouping factor.

If conditional variances are requested, they are returned in the same structure as the conditional modes (random effect estimates/predictions).

Value

A list of matrices with variances in the diagonal and correlation parameters in the off-diagonal — one matrix for each random effects term in the model. Standard deviations are provided as attributes to the matrices.

Author(s)

Rune Haubo B Christensen

Examples

fm1 <- clmm(rating ~ contact + temp + (1|judge), data=wine)
VarCorr(fm1)

[Package ordinal version 2025.12-29 Index]