anova.clm2 {ordinal}	R Documentation

Likelihood ratio test of cumulative link models

Description

Comparison of cumulative link models in likelihood ratio tests. The models may differ by terms in location, scale and nominal formulae, in link, threshold function and random effect structure.

Usage

## S3 method for class 'clm2'
anova(object, ..., test = c("Chisq", "none"))
## S3 method for class 'clmm2'
anova(object, ..., test = c("Chisq", "none"))

Arguments

object
a clm2 object.

...
one or more additional clm2 objects.

test
if test = "none" the p-value for the likelihood ratio test is suppressed.

Value

The method returns an object of class Anova (for printing) and data.frame with the following elements

Model
character description of the cumulative link models being compared. Location, scale and nominal formulae are separated by "|"s in this order.

Resid.df
the residual degrees of freedom

-2logLik
twice the negative log likelihood (proportional to the deviance)

Test
indication of which models are being compared.

DF
the difference in the degrees of freedom in the models being compared, i.e. the degrees of freedom for the chi-squared test.

LR stat.
the likelihood ratio statistic.

Pr(Chi)
the p-value from the likelihood ratio test. Absent if test = "none".

Author(s)

Rune Haubo B Christensen

See Also

clm2, addterm, dropterm and anova.default

Examples

options(contrasts = c("contr.treatment", "contr.poly"))
m1 <- clm2(SURENESS ~ PROD, scale = ~PROD, data = soup,
          link = "logistic")

## anova
anova(m1, update(m1, scale = ~.-PROD))
mN1 <- clm2(SURENESS ~ 1, nominal = ~PROD, data = soup,
           link = "logistic")
anova(m1, mN1)
anova(m1, update(m1, scale = ~.-PROD), mN1)

## Fit model from polr example:
if(require(MASS)) {
    fm1 <- clm2(Sat ~ Infl + Type + Cont, weights = Freq, data = housing)
    anova(fm1, update(fm1, scale =~ Cont))
}

[Package ordinal version 2025.12-29 Index]