clm.control {ordinal}	R Documentation

Set control parameters for cumulative link models

Description

Set control parameters for cumulative link models

Usage

clm.control(method = c("Newton", "model.frame", "design", "ucminf", "nlminb",
   "optim"), 
   sign.location = c("negative", "positive"), 
   sign.nominal = c("positive", "negative"), 
   ..., trace = 0L,
   maxIter = 100L, gradTol = 1e-06, maxLineIter = 15L, relTol = 1e-6,
   tol = sqrt(.Machine$double.eps), maxModIter = 5L,
   convergence = c("warn", "silent", "stop", "message"))

Arguments

method
"Newton" fits the model by maximum likelihood and "model.frame" cause clm to return the model.frame, "design" causes clm to return a list of design matrices etc. that can be used with clm.fit. ucminf, nlminb and optim refer to general purpose optimizers.

sign.location
change sign of the location part of the model.

sign.nominal
change sign of the nominal part of the model.

trace
numerical, if > 0 information is printed about and during the optimization process. Defaults to 0.

maxIter
the maximum number of Newton-Raphson iterations. Defaults to 100.

gradTol
the maximum absolute gradient; defaults to 1e-6.

maxLineIter
the maximum number of step halfings allowed if a Newton(-Raphson) step over shoots. Defaults to 15.

relTol
relative convergence tolerence: relative change in the parameter estimates between Newton iterations. Defaults to 1e-6.

tol
numerical tolerence on eigenvalues to determine negative-definiteness of Hessian. If the Hessian of a model fit is negative definite, the fitting algorithm did not converge. If the Hessian is singular, the fitting algorithm did converge albeit not to a unique optimum, so one or more parameters are not uniquely determined even though the log-likelihood value is.

maxModIter
the maximum allowable number of consecutive iterations where the Newton step needs to be modified to be a decent direction. Defaults to 5.

convergence
action to take if the fitting algorithm did not converge.

...
control arguments parsed on to ucminf, nlminb or optim.

Value

a list of control parameters.

Author(s)

Rune Haubo B Christensen

See Also

clm

[Package ordinal version 2025.12-29 Index]