clm2.control {ordinal}	R Documentation

Set control parameters for cumulative link models

Description

Set control parameters for cumulative link models

Usage

clm2.control(method = c("ucminf", "Newton", "nlminb", "optim",
            "model.frame"), ..., convTol = 1e-4,
            trace = 0, maxIter = 100, gradTol = 1e-5,
            maxLineIter = 10)

Arguments

method
the optimizer used to maximize the likelihood function. "Newton" only works for models without scale, structured thresholds and flexible link functions, but is considerably faster than the other optimizers when applicable. model.frame simply returns a list of model frames with the location, scale and nominal model frames. "optim" uses the "BFGS" method.

...
control arguments passed on to the chosen optimizer; see ucminf, optim, and nlminb for details.

convTol
convergence criterion on the size of the maximum absolute gradient.

trace
numerical, if > 0 information is printed about and during the optimization process. Defaults to 0.

maxIter
the maximum number of Newton-Raphson iterations. Defaults to 100.

gradTol
the maximum absolute gradient. This is the termination criterion and defaults to 1e-5.

maxLineIter
the maximum number of step halfings allowed if a Newton(-Raphson) step over shoots. Defaults to 10.

Value

a list of control parameters.

Author(s)

Rune Haubo B Christensen

See Also

clm2

[Package ordinal version 2025.12-29 Index]