clmm.control {ordinal}	R Documentation

Set control parameters for cumulative link mixed models

Description

Set control parameters for cumulative link mixed models

Usage

clmm.control(method = c("nlminb", "ucminf", "model.frame"), ..., trace = 0,
maxIter = 50, gradTol = 1e-4, maxLineIter = 50,  useMatrix = FALSE,
innerCtrl = c("warnOnly", "noWarn", "giveError"),
checkRanef = c("warn", "error", "message"))

Arguments

method
the optimizer used to maximize the marginal likelihood function.

...
control arguments passed on to the optimizer; see ucminf for details. ucminf for details.

trace
numerical, if > 0 information is printed about and during the outer optimization process, if < 0 information is also printed about the inner optimization process. Defaults to 0.

maxIter
the maximum number of Newton updates of the inner optimization. 50.

gradTol
the maximum absolute gradient of the inner optimization.

maxLineIter
the maximum number of step halfings allowed if a Newton(-Raphson) step over shoots during the inner optimization.

useMatrix
if TRUE, a general implementation of the Laplace approximation using the Matrix package is used, while if FALSE (default), a C implementation of the Laplace approximation valid only for models with a single random effects term is used when possible. TRUE is not valid for models fitted with quadrature methods.

innerCtrl
the use of warnings/errors if the inner optimization fails to converge.

checkRanef
the use of message/warning/error if there are more random effects than observations.

Value

a list of control parameters

Author(s)

Rune Haubo B Christensen

See Also

clmm

[Package ordinal version 2025.12-29 Index]