convergence {ordinal}	R Documentation

Check convergence of cumulative link models

Description

Check the accuracy of the parameter estimates of cumulative link models. The number of correct decimals and number of significant digits is given for the maximum likelihood estimates of the parameters in a cumulative link model fitted with clm.

Usage

convergence(object, ...)

## S3 method for class 'clm'
convergence(object, digits = max(3, getOption("digits") - 3),
   tol = sqrt(.Machine$double.eps), ...)

Arguments

object
for the clm method an object of class "clm", i.e., the result of a call to clm.

digits
the number of digits in the printed table.

tol
numerical tolerence to judge if the Hessian is positive definite from its smallest eigenvalue.

...
arguments to a from methods. Not used by the clm method.

Details

The number of correct decimals is defined as...

The number of significant digits is defined as ...

The number of correct decimals and the number of significant digits are determined from the numerical errors in the parameter estimates. The numerical errors are determined from the Method Independent Error Theorem (Elden et al, 2004) and is based on the Newton step evaluated at convergence.

Value

Convergence information. In particular a table where the Error column gives the numerical error in the parameter estimates. These numbers express how far the parameter estimates in the fitted model are from the true maximum likelihood estimates for this model. The Cor.Dec gives the number of correct decimals with which the the parameters are determined and the Sig.Dig gives the number of significant digits with which the parameters are determined.

The number denoted logLik.error is the error in the value of log-likelihood in the fitted model at the parameter values of that fit. An accurate determination of the log-likelihood is essential for accurate likelihood ratio tests in model comparison.

Author(s)

Rune Haubo B Christensen

References

Elden, L., Wittmeyer-Koch, L. and Nielsen, H. B. (2004) Introduction to Numerical Computation — analysis and Matlab illustrations. Studentliteratur.

Examples

## Simple model:
fm1 <- clm(rating ~ contact + temp, data=wine)
summary(fm1)
convergence(fm1)

[Package ordinal version 2025.12-29 Index]