ordinal-package {ordinal}	R Documentation

Regression Models for Ordinal Data via Cumulative Link (Mixed) Models

Description

This package facilitates analysis of ordinal (ordered categorical data) via cumulative link models (CLMs) and cumulative link mixed models (CLMMs). Robust and efficient computational methods gives speedy and accurate estimation. A wide range of methods for model fits aids the data analysis.

Details

Package: 	 ordinal

Type: 	 Package

License: 	 GPL (>= 2)

LazyLoad: 	 yes

This package implements cumualtive link models and cumulative link models with normally distributed random effects, denoted cumulative link mixed (effects) models. Cumulative link models are also known as ordered regression models, proportional odds models, proportional hazards models for grouped survival times and ordered logit/probit/... models.

Cumulative link models are fitted with clm and the main features are:

A range of standard link functions are available.

In addition to the standard location (additive) effects, scale (multiplicative) effects are also allowed.

nominal effects are allowed for any subset of the predictors — these effects are also known as partial proportional odds effects when using the logit link.

Restrictions can be imposed on the thresholds/cut-points, e.g., symmetry or equidistance.

A (modified) Newton-Raphson algorithm provides the maximum likelihood estimates of the parameters. The estimation scheme is robust, fast and accurate.

Rank-deficient designs are identified and unidentified coefficients exposed in print and summary methods as with glm.

A suite of standard methods are available including anova, add/drop-methods, step, profile, confint.

A slice method facilitates illustration of the likelihood function and a convergence method summarizes the accuracy of the model estimation.

The predict method can predict probabilities, response class-predictions and cumulative probabilities, and it provides standard errors and confidence intervals for the predictions.

Cumulative link mixed models are fitted with clmm and the main features are:

Any number of random effect terms can be included.

The syntax for the model formula resembles that of lmer from the lme4 package.

Nested random effects, crossed random effects and partially nested/crossed random effects are allowed.

Estimation is via maximum likelihood using the Laplace approximation or adaptive Gauss-Hermite quadrature (one random effect).

Vector-valued and correlated random effects such as random slopes (random coefficient models) are fitted with the Laplace approximation.

Estimation employs sparse matrix methods from the Matrix package.

During model fitting a Newton-Raphson algorithm updates the conditional modes of the random effects a large number of times. The likelihood function is optimized with a general purpose optimizer.

A major update of the package in August 2011 introduced new and improved implementations of clm and clmm. The old implementations are available with clm2 and clmm2. At the time of writing there is functionality in clm2 and clmm2 not yet available in clm and clmm. This includes flexible link functions (log-gamma and Aranda-Ordaz links) and a profile method for random effect variance parameters in CLMMs. The new implementations are expected to take over the old implementations at some point, hence the latter will eventually be deprecated and defunct.

Author(s)

Rune Haubo B Christensen

Maintainer: Rune Haubo B Christensen <rune.haubo@gmail.com>

Examples

## A simple cumulative link model:
fm1 <- clm(rating ~ contact + temp, data=wine)
summary(fm1)

## A simple cumulative link mixed model:
fmm1 <- clmm(rating ~ contact + temp + (1|judge), data=wine)
summary(fmm1)

[Package ordinal version 2025.12-29 Index]