soup {ordinal}	R Documentation

Discrimination study of packet soup

Description

The soup data frame has 1847 rows and 13 variables. 185 respondents participated in an A-not A discrimination test with sureness. Before experimentation the respondents were familiarized with the reference product and during experimentation, the respondents were asked to rate samples on an ordered scale with six categories given by combinations of (reference, not reference) and (sure, not sure, guess) from 'referene, sure' = 1 to 'not reference, sure' = 6.

Usage

soup

Format

RESP

factor with 185 levels: the respondents in the study.

PROD

factor with 2 levels: index reference and test products.

PRODID

factor with 6 levels: index reference and the five test product variants.

SURENESS

ordered factor with 6 levels: the respondents ratings of soup samples.

DAY

factor with two levels: experimentation was split over two days.

SOUPTYPE

factor with three levels: the type of soup regularly consumed by the respondent.

SOUPFREQ

factor with 3 levels: the frequency with which the respondent consumes soup.

COLD

factor with two levels: does the respondent have a cold?

EASY

factor with ten levels: How easy did the respondent find the discrimation test? 1 = difficult, 10 = easy.

GENDER

factor with two levels: gender of the respondent.

AGEGROUP

factor with four levels: the age of the respondent.

LOCATION

factor with three levels: three different locations where experimentation took place.

Source

Data are produced by Unilever Research. Permission to publish the data is granted.

References

Christensen, R. H. B., Cleaver, G. and Brockhoff, P. B.(2011) Statistical and Thurstonian models for the A-not A protocol with and without sureness. Food Quality and Preference, 22, pp. 542-549.

[Package ordinal version 2025.12-29 Index]