update.clm2 {ordinal}	R Documentation

Update method for cumulative link models

Description

Update method for cumulative link models fitted with clm2. This makes it possible to use e.g. update(obj, location = ~ . - var1, scale = ~ . + var2)

Usage

## S3 method for class 'clm2'
update(object, formula., location, scale, nominal,...,
        evaluate = TRUE)
## S3 method for class 'clmm2'
update(object, formula., location, scale, nominal,...,
        evaluate = TRUE)

Arguments

object
a clm2 object.

formula.
not used—unfortunately this argument is part of the default method.

location
an optional new formula for the location; see update.formula for details.

scale
an optional new formula for the scale; see update.formula for details.

nominal
an optional new formula for nominal effects; see update.formula for details.

...
additional arguments to the call, or arguments with changed values.

evaluate
if true evaluate the new call else return the call.

Value

If evaluate = TRUE the fitted object is returned, otherwise the updated call.

Author(s)

Rune Haubo B Christensen

Examples

options(contrasts = c("contr.treatment", "contr.poly"))

m1 <-  clm2(SURENESS ~ PROD, scale = ~PROD, data = soup,
           link = "logistic")

m2 <- update(m1, link = "probit")
m3 <- update(m1, link = "cloglog")
m4 <- update(m1, link = "loglog")
anova(m1, update(m1, scale = ~.-PROD))
mT1 <- update(m1, threshold = "symmetric")

[Package ordinal version 2025.12-29 Index]